+++ src/client/Presets/Sweep.hpp
@@ -46,7 +46,7 @@ int SweepPreset(EnvVars&          ev,
 {
   if (TransferBench::GetNumRanks() > 1) {
     Utils::Print("[ERROR] Sweep preset currently not supported for multi-node\n");
-    return 1;
+    return ERR_FATAL;
   }
 
   bool const isRandom = (presetName == "rsweep");
@@ -104,33 +104,33 @@ int SweepPreset(EnvVars&          ev,
   for (auto ch : sweepSrc) {
     if (!strchr(MemTypeStr, ch)) {
       printf("[ERROR] Unrecognized memory type '%c' specified for sweep source\n", ch);
-      return 1;
+      return ERR_FATAL;
     }
     if (strchr(sweepSrc.c_str(), ch) != strrchr(sweepSrc.c_str(), ch)) {
       printf("[ERROR] Duplicate memory type '%c' specified for sweep source\n", ch);
-      return 1;
+      return ERR_FATAL;
     }
   }
 
   for (auto ch : sweepDst) {
     if (!strchr(MemTypeStr, ch)) {
       printf("[ERROR] Unrecognized memory type '%c' specified for sweep destination\n", ch);
-      return 1;
+      return ERR_FATAL;
     }
     if (strchr(sweepDst.c_str(), ch) != strrchr(sweepDst.c_str(), ch)) {
       printf("[ERROR] Duplicate memory type '%c' specified for sweep destination\n", ch);
-      return 1;
+      return ERR_FATAL;
     }
   }
 
   for (auto ch : sweepExe) {
     if (!strchr(ExeTypeStr, ch)) {
       printf("[ERROR] Unrecognized executor type '%c' specified for sweep executor\n", ch);
-      return 1;
+      return ERR_FATAL;
     }
     if (strchr(sweepExe.c_str(), ch) != strrchr(sweepExe.c_str(), ch)) {
       printf("[ERROR] Duplicate executor type '%c' specified for sweep executor\n", ch);
-      return 1;
+      return ERR_FATAL;
     }
   }
 
@@ -340,7 +340,7 @@ int SweepPreset(EnvVars&          ev,
 
     if (!TransferBench::RunTransfers(cfg, transfers, results)) {
       Utils::PrintErrors(results.errResults);
-      if (!continueOnErr) return 1;
+      if (!continueOnErr) return ERR_FATAL;
     } else {
       Utils::PrintResults(ev, numTestsRun, transfers, results);
     }
@@ -372,5 +372,5 @@ int SweepPreset(EnvVars&          ev,
     }
   }
   if (fp) fclose(fp);
-  return 0;
+  return ERR_NONE;
 }